+++ CHANGELOG.md
@@ -10,8 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking Change
 
--   Adds user context to all functions exposed to the user, and to API and Recipe interface functions.
+-   Adds user context to all functions exposed to the user, and to API and Recipe interface functions. This is a non breaking change for User exposed function calls, but a breaking change if you are using the Recipe or APIs override feature
 -   Returns session from API interface functions that create a session
+-   Renames functions in ThirdPartyEmailPassword recipe (https://github.com/supertokens/supertokens-node/issues/219):
+    -   Recipe Interface:
+        -   `signInUp` -> `thirdPartySignInUp`
+        -   `signUp` -> `emailPasswordSignUp`
+        -   `signIn` -> `emailPasswordSignIn`
+    -   API Interface:
+        -   `emailExistsGET` -> `emailPasswordEmailExistsGET`
+    -   User exposed functions (in `recipe/thirdpartyemailpassword/index.ts`)
+        -   `signInUp` -> `thirdPartySignInUp`
+        -   `signUp` -> `emailPasswordSignUp`
+        -   `signIn` -> `emailPasswordSignIn`
 
 ### Change: